rename createSecret to updateSecret
