+++ src/pages/namespace/Settings/Secrets/CreateEdit.tsx
@@ -17,7 +17,7 @@ import Button from "~/design/Button";
 import FormErrors from "~/componentsNext/FormErrors";
 import Input from "~/design/Input";
 import { Textarea } from "~/design/TextArea";
-import { useCreateSecret } from "~/api/secrets/mutate/createSecret";
+import { useCreateSecret } from "~/api/secrets/mutate/updateSecret";
 import { useTranslation } from "react-i18next";
 import { zodResolver } from "@hookform/resolvers/zod";
